Added all files
Data folder has all the required data to load the files into the app.

microbiome_explorer is the django created folder for the app.

rdp files folder consists of all the files required to get rdp results
and parse the rdp output.

Read_&_sample folder consists of the scripts to obtain the information
about the reads and samples. Also consists of the output files of the
same.

Taxonomy folder consists of all files to obtain the taxa table.

Variables folder consists of files to obtain the sample variables table.